Finalize support for Triton
Triton installation is a bit tricky as it's mandatory to have
either NVIDIA or AMD backend installed on the system (which can
also be different from version-to-version and be hardware dependent.

So here no automatic installation is provided with the commit. Instead
PYTORCH_INDEX env variable is introduced to change default CPU pytorch
nightly wheels to whatever a user desires.